requirements.txt
torch
torchvision
numpy
matplotlib
streamlit
pillow
scikit-learn
seaborn
tqdm 